remove old packettype + fix: apply routing filter BEFORE delaying
