+++ common/nymsphinx/framing/src/codec.rs
@@ -85,8 +85,6 @@ impl Decoder for NymCodec {
             match header.packet_type {
                 PacketType::Outfox => NymPacket::outfox_from_bytes(slice)?,
                 PacketType::Mix => NymPacket::sphinx_from_bytes(slice)?,
-                #[allow(deprecated)]
-                PacketType::Vpn => NymPacket::sphinx_from_bytes(slice)?,
             }
         } else {
             return Ok(None);
