+++ nym-node/src/node/replay_protection/mod.rs
@@ -1,0 +1,2 @@
+// Copyright 2025 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: GPL-3.0-only
